+++ R/domin.r
@@ -40,12 +40,14 @@
 #' Typical usage of \code{consmodel} is to pass "1" to set the intercept as the baseline and control for its value when the baseline model's fit statistic value is not 0 (e.g., if using the AIC or BIC as a fit statistic; see examples).
 #' 
 #' As such, this vector is used to set a baseline for the fit statistic when it is non-0. 
-#' @param reverse Logical. If \code{TRUE} then standardized vector, ranks, and complete dominance Designations are reversed in their interpretation.  
+#' @param reverse Logical. If \code{TRUE} then standardized vector, ranks, and complete dominance designations are reversed in their interpretation.  
 #' 
 #' This argument should be changed to \code{TRUE} if the fit statistic used decreases with better fit to the data (e.g., AIC, BIC). 
 #' @param ... Additional arguments passed to the function call in the \code{reg} argument.
 #'
 #' @section Notes:
+#' 
+#' `domin` is an R port of the Stata command with the same name (see Luchman, 2021).
 #' 
 #' `domin` has been superseded by [`domir`].
 #'
@@ -86,7 +88,7 @@
 #' \item x1, x2, x3, x4
 #' }
 #' 
-#' \code{domin} parses \code{formula_overall} to obtain all the terms in it and combines them with \code{sets}.  When parsing \code{formula_overall}, only the processing that is available in the \code{stats} package is applied.  Note that \code{domin} is not programmed to process terms of order > 1 (i.e., interactions/products) appropriately (i.e., only include in the presence of lower order component terms).
+#' \code{domin} parses \code{formula_overall} to obtain all the terms in it and combines them with \code{sets}.  When parsing \code{formula_overall}, only the processing that is available in the \code{stats} package is applied.  Note that \code{domin} is not programmed to process terms of order > 1 (i.e., interactions/products) appropriately (i.e., only include in the presence of lower order component terms). \code{domin} also does not allow \code{offset} terms.
 #' 
 #' From these combinations, the predictive models are constructed and called. The predictive model call includes the entries in \code{all}, applies the appropriate formula, and reconstructs the function itself. The seven combinations above imply the following series of predictive model calls:
 #' 
@@ -118,7 +120,12 @@
 #' 
 #' In the case that the model object returned by \code{reg} includes its own fit statistic without the need for an extractor function, the user can apply an anonymous function following the required format to extract it.
 #' 
-#' @export
+#' @references 
+#' 
+#' Luchman, J. N. (2021). Relative importance analysis in Stata using dominance analysis: domin and domme. 
+#' The Stata Journal, 21, 2. doi: 10.1177/1536867X211025837.
+#' 
+#' @export 
 #' @examples
 #' ## Basic linear model with r-square
 #' 
@@ -176,6 +183,9 @@ domin <-
     
     if (!inherits(formula_overall, "formula")) 
       stop(paste(formula_overall, "is not a 'formula' class object."))
+    
+    if (!is.null(attr(stats::terms(formula_overall), "offset"))) 
+      stop("'offset()' terms not allowed in formula object.")
     
     if (!is.list(fitstat)) 
       stop("fitstat is not a list.")