+++ R/domir.r
@@ -12,28 +12,28 @@
 #' @param .obj A [`formula`], [`Formula`][Formula::Formula], or [`list`]. 
 #' 
 #' Parsed to produce subsets of elements to submit to `.fct`. Always submits 
-#' subsets of `.obj` the same type to `.fct` as the same type or class as 
-#' the input object.
+#' subsets of `.obj` that are of the same [`class`] to `.fct` and are always 
+#' submitted as the first, unnamed argument.
 #' 
-#' @param .fct A function/closure or string function name.
+#' @param .fct A [`function`]/closure or string function name.
 #' 
 #' Applied to all subsets of elements as received from `.obj`.  
 #' Must return a length 1 (scalar), numeric, atomic vector.
 #' 
 #' @param .set A `list`.
 #' 
-#' Must be comprised of elements of the same type or class as `.obj`.  
+#' Must be comprised of elements of the same class as `.obj`.  
 #' Elements of the list can be named.
 #'
 #' @param .wst Not yet used.
 #' 
 #' @param .all A `formula`, `Formula`, or `list`. 
 #' 
-#' Must be the same type or class as `.obj`.
+#' Must be the same class as `.obj`.
 #' 
 #' @param .adj A `formula`, `Formula`, or `list`.
 #' 
-#' Must be the same type or class as `.obj`.
+#' Must be the same class as `.obj`.
 #' 
 #' @param .cdl Logical.  
 #' 
@@ -50,7 +50,7 @@
 #' designations are reversed in their interpretation.  
 #' 
 #' @param ... Passes arguments to other methods; passes arguments to 
-#' function in `.fct`.
+#' the function in `.fct`.
 #'
 #' @return Returns an object of [`class`] "domir" which is a composed of the 
 #' following elements:
@@ -83,7 +83,7 @@
 #' 
 #' `.obj`s elements are parsed and used to determine the required number of 
 #' subsets included the dominance analysis.  How the elements are parsed is 
-#' determined depends on `.obj`s type or class method.
+#' determined depends on `.obj`s class method.
 #' 
 #' ### `formula`
 #' 
@@ -147,23 +147,23 @@
 #' values for all subsets as well as from `Value_All`.
 #' 
 #' The `formula` method for `.adj` does not allow a left hand side but 
-#' allows constants/intercepts (i.e., `~ 1`) to be included as a valid 
+#' allows the intercept (i.e., `~ 1`) to be included as a valid 
 #' element in the right hand size even when not explicitly included in `.obj`.
 #' 
 #' ### Additional Details
 #' 
 #' All element parsing methods for `domir` will submit subsets generated as an 
-#' object of the same type as `.obj` (i.e., a `formula` in `.obj` will be 
+#' object of the same class as `.obj` (i.e., a `formula` in `.obj` will be 
 #' submitted as a `formula`) to the `.fct` as the first, unnamed argument.
 #' 
 #' ## `.fct` as Analysis Pipeline
 #' 
-#' The function in `.fct` will be called repeatedly, once for each subset 
+#' The function in `.fct` will be called repeatedly; once for each subset 
 #' of elements created from `.obj`.
 #' 
 #' `.fct` is expected to be a complete analysis pipeline that receives a 
 #' subset of elements from `.obj`, uses the subset of elements from `.obj` in 
-#' the type/class as received to generate a predictive model, and 
+#' the class as received to generate a predictive model, and 
 #' extracts a returned value of the appropriate type to dominance analyze.  
 #' 
 #' At current, only atomic (i.e., non-`list`), numeric scalars (i.e., 
@@ -174,6 +174,22 @@
 #' output adhere to these requirements.  In most circumstances, 
 #' the user will have to make their own named or anonymous function to 
 #' supply to `.fct` to meet `domir`s requirements.
+#' 
+#' # Notes
+#' 
+#' ## `formula` method
+#' 
+#' `formula` objects and all arguments for the `formula` method do not allow 
+#' [`offset`] terms. When the user needs to include offset terms in their model, 
+#' they could be included by the user in the `.fct` as an 
+#' [`update`][update.formula] to the submitted `formula` object or through 
+#' other similar means.
+#' 
+#' `formula`s with second-order or higher terms (i.e., interactions like
+#'  `~ a*b`) are not, by default, used differently from first-order terms. 
+#'  The dominance analysis results from models using such terms may not produce 
+#'  useful results unless the user ensures that second-order and higher terms 
+#'  are used appropriately in `.fct`.
 #' 
 #' @export
 #' @examples
@@ -222,6 +238,7 @@
 #'   .set = list(trns = ~ am + gear, 
 #'     eng = ~ cyl + vs, misc = ~ qsec + drat))
 #'   
+#'   
 #' ## Linear model using AIC
 #' 
 #' lm_aic <- function(fml, data) { 
@@ -256,7 +273,7 @@ domir.formula <- function(
   }
   
   # Process formula ----
-  # obtain rhs name vector from `.obj`
+  # Obtain 'right hand side'/RHS name vector from `.obj`
   RHS_names <- 
     attr(stats::terms(.obj), "term.labels") 
   
@@ -266,63 +283,97 @@ domir.formula <- function(
       attr(stats::terms(.obj), "intercept") 
     ) 
   
-  # Obtain lhs (if included)
+  # Offsets are not allowed - too complex to accommodate
+  if (!is.null(attr(stats::terms(.obj), "offset"))) {
+    
+    stop("'offset()' terms are not allowed in formula-class '.obj's.\n", 
+         "Offset terms can be included using 'update()' in the call to .fct.\n", 
+         "For example, 'function(fml) {fml_ofst <- update(fml, . ~ . + offset(x)); ...}'.", 
+         call. = FALSE)
+    
+  }
+
+  
+  # Obtain 'left hand side'/LHS (if included)
   if (attr(stats::terms(.obj), "response") == 1)
     LHS_names <- 
-    attr(stats::terms(.obj), "variables")[[
-      attr(stats::terms(.obj), "response") + 1
+    rownames(attr(stats::terms(.obj), "factors"))[[
+      attr(stats::terms(.obj), "response")
     ]]
   
   else LHS_names <- NULL
-
+  
+  # Define formula argument checks function ----
+  fml_check <- function(fml, name, chk_int) {
+    
+    # enforce formula class
+    if (!inherits(fml, "formula")) {
+      
+      stop(name, " must be a 'formula'.", call. = FALSE)
+      
+    }
+    
+    # enforce no response/lhs
+    if (attr(stats::terms(fml), "response") == 1) {
+      
+      stop(name, " must not have a response/left hand side.", call. = FALSE)
+      
+    }
+    
+    # enforce no offset term
+    if (!is.null(attr(stats::terms(fml), "offset"))) {
+      
+      stop("'offset()' terms not allowed in ", name, ".", 
+           call. = FALSE)
+      
+    }
+    
+    # enforce no empty model - when applicable
+    if ((length(attr(stats::terms(fml), "term.labels")) == 0) && chk_int) {
+      
+      stop(name, " cannot be an intercept-only/empty formula.", 
+           call. = FALSE)
+      
+    }
+    
+  }
+  
+  # Define function to remove rhs names
+  rhs_name_remover <- function(obj, rhs, name) {
+    
+    # find locations of names to remove (if any)
+    remove_loc <- 
+      unlist( lapply(obj, 
+               function(x) which(rhs %in% x)) )
+    
+    # error if some of the names did not match 
+    if (length(remove_loc) != length(unlist(obj))) {
+      
+      wrong_terms <- 
+        paste( unlist(obj)[ which(!(unlist(obj) %in% rhs)) ], 
+          collapse = " ")
+      
+      stop("Names ", wrong_terms, " in ", name, 
+           " do not match any names in '.obj'.", 
+           call. = FALSE)
+    }
+    
+    return(rhs[-remove_loc])
+    
+  }
   
   # Process '.all' ----
   if (!is.null(.all)) {
     
-    if (!inherits(.all, "formula")) {
-      
-      stop("'.all' must be a 'formula'.", call. = FALSE)
-      
-    }
-    
-    if (attr(stats::terms(.all), "response") == 1) {
-      
-      stop("'.all' must not have a response/left hand side.", call. = FALSE)
-      
-    }
-    
-    if (length(attr(stats::terms(.all), "term.labels")) == 0) {
-      
-      stop("'.all' cannot be an intercept-only model.", 
-           call. = FALSE)
-      
-    }
+    # apply checks to '.all' argument
+    fml_check(.all, "'.all'", TRUE)
     
     # name vector from '.all'
     All_names <- attr(stats::terms(.all), "term.labels")
     
-    # remove names from `RHS_names` list if in `.all`
-    all_remove_loc <- 
-      unlist(
-        lapply(All_names, 
-                    function(x) which(RHS_names %in% x))
-        )
-    
-    if (length(all_remove_loc) != length(unlist(All_names))) {
-      
-      wrong_all_terms <- 
-        paste(
-          unlist(All_names)[
-            which(!(unlist(All_names) %in% 
-                      RHS_names))], 
-          collapse = " ")
-      
-      stop("Names ", wrong_all_terms,
-           " in '.all' do not match any names in '.obj'.", 
-           call. = FALSE)
-    }
-    
-    RHS_names <- RHS_names[-all_remove_loc]
+    # check names in '.all' and remove from rhs
+    RHS_names <- 
+      rhs_name_remover(All_names, RHS_names, "'.all'")
     
   }
   
@@ -331,52 +382,35 @@ domir.formula <- function(
   # Process '.adj' ----
   if (!is.null(.adj)) {
     
-    if (!inherits(.adj, "formula")) {
-      
-      stop("'.adj' must be a 'formula'.", call. = FALSE)
-      
-    }
+    # apply checks to '.adj' argument
+    fml_check(.adj, "'.adj'", FALSE)
     
-    if (attr(stats::terms(.adj), "response") == 1) {
-      
-      stop("'.adj' must not have a response/left hand side.", call. = FALSE)
-      
-    }
+    # if '.adj' is an intercept-only model - checks/include
+    if (
+      (length(attr(stats::terms(.adj), "term.labels")) == 0) && 
+      (attr(stats::terms(.adj), "intercept") == 1) && Intercept
+    ) Adj_names <- "1"
     
-    # if .adj is an intercept-only model
-    if (length(attr(stats::terms(.adj), "term.labels")) == 0) {
-      
-      Adj_names <- "1"
-      
-    }
+    else if (
+      (length(attr(stats::terms(.adj), "term.labels")) == 0) && 
+      (attr(stats::terms(.adj), "intercept") == 1) && !Intercept
+    ) stop("'.adj' cannot add an intercept when it is removed \n", 
+    "from the formula in '.obj'.", 
+           call. = FALSE)
     
+    else if (
+      (length(attr(stats::terms(.adj), "term.labels")) == 0) && 
+      (attr(stats::terms(.adj), "intercept") == 0)
+    ) stop("'.adj' cannot be an empty formula.", call. = FALSE)
+
     else {
       
       # name vector in '.adj'
       Adj_names <- attr(stats::terms(.adj), "term.labels")
       
-      # remove names from `RHS_names` list if in `.adj`
-      adj_remove_loc <- 
-        unlist(
-          lapply(Adj_names, 
-                      function(x) which(RHS_names %in% x))
-          )
-      
-      if (length(adj_remove_loc) != length(unlist(Adj_names))) {
-        
-        wrong_adj_terms <- 
-          paste(
-            unlist(Adj_names)[
-              which(!(unlist(Adj_names) %in% 
-                        RHS_names))], 
-            collapse = " ")
-        
-        stop("Names ", wrong_adj_terms,
-             " in '.adj' do not match any names in '.obj'.", 
-             call. = FALSE)
-      }
-      
-      RHS_names <- RHS_names[-adj_remove_loc]
+      # check names in '.adj' and remove from rhs
+      RHS_names <- 
+        rhs_name_remover(Adj_names, RHS_names, "'.adj'")
       
     }
     
@@ -391,77 +425,28 @@ domir.formula <- function(
       stop("'.set' must be a 'list'.", call. = FALSE)
     }
     
-    if (!all(sapply(.set, inherits, "formula"))) {
-      
-      elements_not_formula <- 
-        paste(which((!sapply(.set, inherits, "formula"))), 
-              collapse = " ")
-      
-      stop("Each element of list in '.set' must be a 'formula'.\n", 
-           "Elements ", elements_not_formula, 
-           " are not a 'formula'.", call. = FALSE)
-    }
-    
-    if (any(sapply(.set, function(x) 
-      attr(stats::terms(x), "response") == 1))) {
-      
-      elements_has_response <- 
-        paste(which(sapply(.set, 
-                           function(x) attr(stats::terms(x), "response")) > 0 ), 
-              collapse = " ")
-      
-      stop("Elements in '.set' must not have responses/left hand sides.\n",
-           "Elements ", elements_has_response, " have responses.", 
-           call. = FALSE)
-      
-    }
-    
-    if (any(sapply(.set, function(x) 
-      length(attr(stats::terms(x), "term.labels")) == 0))) {
-      
-      elements_intercept_only <- 
-        paste(
-          which(sapply(.set, 
-                       function(x) 
-                         length(attr(stats::terms(x), "term.labels"))) > 0 ), 
-          collapse = " ")
-      
-      stop("Elements in '.set' must not be intercepts-only.\n",
-           "Elements ", elements_intercept_only, " have no terms.", call. = FALSE)
-      
-    }
+    # apply checks to all elements of '.set' argument
+    .mapply(fml_check, list(.set, 
+            paste0("Position ", 1:(length(.set)), " of '.set'."), 
+            rep(TRUE, times = length(.set))),
+            NULL)
     
     # name vectors from each element of '.set'
     Set_names <- 
       lapply(.set, 
              function(x) attr(stats::terms(x), "term.labels"))
     
-    # remove IVs from `RHS_names` list if in `.set`
-    set_remove_loc <- 
-      unlist(
-        lapply(Set_names, 
-                    function(x) which(RHS_names %in% x))
-        )
-    
-    if (length(set_remove_loc) != length(unlist(Set_names))) {
+    # recursively check names in '.set'-s and remove from rhs
+    for (set in 1:length(Set_names)) {
       
-      wrong_set_terms <- 
-        paste(
-          unlist(Set_names)[
-            which(!(unlist(Set_names) %in% 
-                      RHS_names))], 
-          collapse = " ")
+      RHS_names <- 
+        rhs_name_remover(Set_names[[set]], RHS_names, 
+                         paste0("Position ", set, " of '.set'."))
       
-      stop("Names ", wrong_set_terms,
-           " in '.set' do not match any names ", 
-           "in .'obj'.", 
-           call. = FALSE)
     }
     
-    RHS_names <- RHS_names[-set_remove_loc]
-    
     # apply labels to '.set'
-    if (!is.null(names(.set)))    
+    if (!is.null(names(.set))) 
       Set_labels <- names(.set)
     
     else Set_labels <- paste0("set", 1:length(.set))
@@ -491,7 +476,6 @@ domir.formula <- function(
          call. = FALSE)
   
   # Define meta-function to coordinate .fct calls ----
-  
   meta_domir <- 
     function(Selector_lgl, 
              RHS, LHS, 
@@ -541,10 +525,9 @@ domir.formula <- function(
          call. = FALSE)
   
   # Define arguments to `dominance_scalar` ----
-  
   args_list <-  
     list(RHS = append(RHS_names, Set_names), 
-         LHS = deparse(LHS_names),
+         LHS = LHS_names,
          .fct = .fct,
          .all = All_names, 
          .adj = Adj_names, 
@@ -553,7 +536,7 @@ domir.formula <- function(
   
   cons_args <-
     list(RHS = RHS_names, 
-         LHS = deparse(LHS_names),
+         LHS = LHS_names,
          .fct = .fct,
          .all = NULL, 
          .adj = Adj_names, 
@@ -649,15 +632,14 @@ domir.list <- function(...) {
 #' following results:
 #' \itemize{
 #'  \item{Value when all elements are included in `obj`.}
-#'  \item{Value for the elements included in `.all` if any.}  
-#'  \item{Value for the elements included in `.adj` if any.}
+#'  \item{Value for the elements included in `.all`, if any.}  
+#'  \item{Value for the elements included in `.adj`, if any.}
 #'  \item{Matrix describing general dominance values, standardized 
 #'  general dominance values, and the ranking of the general 
 #'  dominance values.}
-#'  \item{Matrix describing the conditional dominance values; when computed}
-#'  \item{Matrix describing the complete dominance designations; when 
-#'  determined}
-#'  \item{If following `summary.domir`, matrix describing the strongest 
+#'  \item{Matrix describing the conditional dominance values, if computed}
+#'  \item{Matrix describing the complete dominance designations, if evaluated}
+#'  \item{If following [`summary.domir`], matrix describing the strongest 
 #'  dominance designations between all elements.}}
 #'  
 #'  The `domir` print method alters dimension names for readability and they 
@@ -748,9 +730,9 @@ print.domir <- function(x, ...) {
 #'  designation between the two elements.}
 #' }
 #' 
-#' @details The summary method for class `domir` is used for obtaining the 
-#' strongest dominance designations (i.e., general, conditional, or complete) 
-#' among the elements.
+#' @details The summary method for class `domir` objects is used for obtaining 
+#' the strongest dominance designations (i.e., general, conditional, or 
+#' complete) among all pairs of dominance analyzed elements.
 #' 
 #' @exportS3Method
 
@@ -786,7 +768,7 @@ summary.domir <- function(object, ...) {
             pairs[2, location] <- 
               ifelse(object$Complete_Dominance[IV1, IV2], 
                      "completely dominates",
-                     "is Completely dominated by")
+                     "is completely dominated by")
             
             next